+++ lldb/unittests/DIL/Runner.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runner.h"
+#include "Runner.h"
 
 #include <fstream>
 #include <iostream>
